+++ CHANGES.md
@@ -16,8 +16,8 @@
 ### 🎉 1.0.0 - 2019-08-??
 
 * *Plugin launch. Everything's new!*
-* New: 5 Free GiveWP extensions supported/ integrated
-* New: 14 Premium GiveWP extensions supported/ integrated
+* New: 6 Free GiveWP extensions supported/ integrated
+* New: 15 Premium GiveWP extensions supported/ integrated
 
 
 ### ⚡ 0.9.0 - 2019-08-06